Add Artifacts to an SBOM (#27)
Resolves #2 . Resolves #22 .

All artifacts (including lazy artifacts) are included in SBOM generation. The artifact versions are resolved to target the platform that Julia is running on. As a happy by-product of this, the version of julia that was used to generate the SBOM is now documented. An advanced user could change the target platform to another OS and CPU architecture, but this has not been tested.