Fill in Declared License field in SBOM (#32)
Resolves #7 [Fill in Declared License field in SBOM]

Adds LicenseCheck.jl to the dependencies and uses it to scan each package and artifact for files that contain recognized software licenses. Makes a best effort to determine which file contains the license file that governs the entire package. Fills in the SBOM with this information.

Added logging to record which files contained license information. Set logging level to -50 to see the declared license file for each package and set to -75 to see every file that contained a software license.

Added a docstring for generateSPDX()